requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
WTForms==3.1.2
email-validator==2.2.0
python-dotenv==1.0.1
APScheduler==3.10.4
gunicorn==22.0.0
